04 auth

+++ controllers/contacts-controllers.js
@@ -5,15 +5,23 @@ const { HttpError, ctrlWrapper } = require("../utils");
 
 // ------------------------------------------------------------------------ | ROUTERS
 const getList = async (req, res) => {
-  const result = await Contact.find({});
-  res.json(result);
+  const { _id: owner } = req.user;
+  const { page = 1, limit = 10 } = req.query;
+  const skip = (page - 1) * limit;
+  const result = await Contact.find({owner},"",{skip, limit}).populate("owner", "email"); // populate() бере ід записане в цьому полі, іде в колекцію яка вказана в contactSchema.owner.ref і записує цей обєкт(це назив поширений запит!)! Другий параметр перелік полів який нам потрібен!
+  const favoriteResult = result.filter(it => it.favorite);
+  const chekingResult = Object.keys(req.query).find(it => it === "favorite");
+  
+  res.json(chekingResult ? favoriteResult : result);
 };
 
 const getById = async (req, res) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
   // const result = await Contact.findOne({ _id: contactId }); - один із варіантів!
 
   const result = await Contact.findById(contactId);
+  if (JSON.stringify(owner) !== JSON.stringify(result.owner)) throw HttpError(404, "Not found");
   if (!result) {
     throw HttpError(404, "Not found");
   }
@@ -21,17 +29,21 @@ const getById = async (req, res) => {
 };
 
 const add = async (req, res) => {
-  const result = await Contact.create(req.body);
+  const { _id: owner } = req.user; // забираємо з реквеста(інфу про якого ми записали в authenticat) і перейм. в owner! 
+  const result = await Contact.create({ ...req.body, owner });
   res.status(201).json(result);
 };
 
 const update = async (req, res) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
 
   if (Object.keys(req.body).length === 0) {
     throw HttpError(400, "missing fields");
   }
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {new:true});
+  const allResult = await Contact.findById(contactId);
+  if (JSON.stringify(owner) !== JSON.stringify(allResult.owner)) throw HttpError(404, "Not found");
+  const result = await Contact.findByIdAndUpdate(contactId, req.body, { new: true });
   if (!result) {
     throw HttpError(404, "Not found");
   }
@@ -39,11 +51,14 @@ const update = async (req, res) => {
 };
 
 const updateFavorite = async (req, res) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
 
   if (Object.keys(req.body).length === 0) {
     throw HttpError(400, "missing field favorite");
   }
+  const allResult = await Contact.findById(contactId);
+  if (JSON.stringify(owner) !== JSON.stringify(allResult.owner)) throw HttpError(404, "Not found");
   const result = await Contact.findByIdAndUpdate(contactId, req.body, {new:true});
   if (!result) {
     throw HttpError(404, "Not found");
@@ -52,7 +67,12 @@ const updateFavorite = async (req, res) => {
 };
 
 const remove = async (req, res) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
+
+  
+  const allResult = await Contact.findById(contactId);
+  if (JSON.stringify(owner) !== JSON.stringify(allResult.owner)) throw HttpError(404, "Not found");
   const result = await Contact.findByIdAndDelete(contactId);
   if (!result) {
     throw HttpError(404, "Not found");
